finish lecture 6
