+++ CS193p_Memorize/CS193p_Memorize/EmojiMemorizeGameView.swift
@@ -12,6 +12,8 @@ struct EmojiMemorizeGameView: View {
     //view model pointer
     @ObservedObject var viewModel: EmojiMemorizeGame
     
+    private let aspectRatio: CGFloat = 2/3
+    
     // main/body view
     var body: some View {
         VStack{
@@ -50,48 +52,15 @@ struct EmojiMemorizeGameView: View {
     }
     
     //card grid view
-    @ViewBuilder
-    var cards : some View {
-        let aspectRadio: CGFloat = 2/3
-        GeometryReader { geometry in
-            let gridItemSize: CGFloat = gridItemWidthThatFits(
-                count: viewModel.cards.count,
-                size: geometry.size,
-                atAspectRatio: aspectRadio
-            )
-            LazyVGrid(columns: [GridItem(.adaptive(minimum: gridItemSize), spacing: 0)], spacing: 0)  {
-                ForEach(viewModel.cards){ card in
-                    CardView(card, themeColor: viewModel.themeColor)
-                        .aspectRatio(aspectRadio, contentMode: .fit)
-                        .padding(4)
-                        .onTapGesture {
-                            viewModel.choose(card)
-                        }
+    private var cards : some View {
+        AspectVGrid( viewModel.cards, aspectRatio: aspectRatio){ card in
+            return CardView(card, themeColor: viewModel.themeColor)
+                .padding(4)
+                .onTapGesture {
+                    viewModel.choose(card)
                 }
             }
         }
-    }
-    
-    func gridItemWidthThatFits(
-        count: Int,
-        size: CGSize,
-        atAspectRatio aspectRatio: CGFloat
-    ) -> CGFloat {
-        let count = CGFloat(count)
-        var columnCount = 1.0
-        repeat {
-            let width = size.width / columnCount
-            let height = width / aspectRatio
-            let rowCount = (count / columnCount).rounded(.up)
-            
-            if rowCount * height < size.height {
-                return (size.width / columnCount).rounded(.down)
-            }
-            
-            columnCount += 1
-        } while columnCount < count
-        return min(size.width / count, size.height * aspectRatio).rounded(.down)
-    }
     
     struct CardView: View {
